// src/api/generated/models/DocTemplate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DocTemplate = {
  id?: number;
  name?: string;
  description?: string;
  tags?: string;
  user_id?: string;
  created_at?: number;
  updated_at?: number;
  is_deleted?: boolean;
  type?: number;
};