// src/api/generated/models/DocTemplateUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DocTemplateUpdate = {
  id?: number;
  name?: string;
  description?: string;
  template_data?: string;
  tags?: string;
  user_id?: number;
  is_deleted?: boolean;
  type?: number;
};